+++ src/containers/Activities/activity.tsx
@@ -1194,7 +1194,7 @@ export default function Activites() {
 
   return (
     <CyDSafeAreaView className='bg-n0 flex-1' edges={['top']}>
-      <PageHeader title={t('ACTIVITIES')} navigation={navigation} />
+      <PageHeader title={'ACTIVITIES'} navigation={navigation} />
       <CyDView className='bg-n20 flex-1'>
         {activityContext.state.activityObjects.length === 0 ||
         activities.length === 0 ? (
